Test for varying styles in chapters
Will need to keep track of chapter and book information.
These things are serialized differently for the first chapter in a book.
Create two test html files.  Run the unit tests on both serialization
styles.
Add outlines of the tests for the new functionality. Will serve as
reminder for when I get back to this project.